# docs/static/swagger/openApiConfig.yaml
openapi: 3.0.3
info:
  title: ''
  version: ''
  description: >
    This is a simple CRUD API application made with Express and documented with
    Swagger
  termsOfService: http://swagger.io/terms/
  contact:
    email: bourdakos1@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  x-github: https://github.com/bee-travels
servers:
  - url: http://localhost:3001/api
    description: Development server server (uses local data)
  - url: https://sandbox-api.example.com:8443/v1
    description: Sandbox server (uses test data)
...
paths:
...
  /business-users:
    get:
      tags:
        - BusinessUsers
      summary: Get a list of all business users
      description: Returns a list of all business users.
      operationId: getAllBusinessUsers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessUser'
        '400':
          description: Error occurred while obtaining the business user list
          content:
            application/json:
              schema:
                type: object
    post:
      tags:
        - BusinessUsers
      summary: Create a new business user
      description: Creates a new business user and links it to a business.
      operationId: createBusinessUser
      requestBody:
        description: The data to create a new business user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessUserRequest'
        required: true
      responses:
        '201':
          description: >-
            Business user created successfully. Returns the created business
            user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessUser'
        '400':
          description: Error creating business user.
...